mem/cast
* lots of memorization and casting clean up. Situations where you'd see
the spell header bar for a level of spells in combat when you had cast
all those and hid them shouldn't happen anymore.  Also fixed the bug for
items with uses headers. Basically in no situation now in "combat"
should you see a header for a group when there is nothing in it.