item/npc
* items now also have a new field, combat effects. This the same
"effect" string that you use to place on npc/pcs in the combat tracker.
This will be applied to the npc/pc that has this item equipped when
placed into the combat tracker. Useful for resists like fire/cold and
regen. Do not use this for persistant effects like abilities and saves,
use the configurable version of those.

* npcs now have an "effect" string that will be applied to them when
dropped into the combat tracker. All you need to do now is know the
string, not the cryptic action text!
* also, items with effects equipped on NPCs placed into the combat
tracker will also be added!